+++ api/src/main/java/net/jqwik/api/constraints/Positive.java
@@ -12,12 +12,12 @@ import java.lang.annotation.*;
  */
 @Target({ ElementType.ANNOTATION_TYPE, ElementType.PARAMETER, ElementType.TYPE_USE })
 @Retention(RetentionPolicy.RUNTIME)
-@ShortRange(min = 0, max = Short.MAX_VALUE)
-@ByteRange(min = 0, max = Byte.MAX_VALUE)
-@IntRange(min = 0, max = Integer.MAX_VALUE)
-@LongRange(min = 0L, max = Long.MAX_VALUE)
-@FloatRange(min = 0f, max = Float.MAX_VALUE)
-@DoubleRange(min = 0, max = Double.MAX_VALUE)
+@ShortRange(min = 0)
+@ByteRange(min = 0)
+@IntRange(min = 0)
+@LongRange(min = 0L)
+@FloatRange(min = 0f)
+@DoubleRange(min = 0)
 @BigRange(min = "0")
 @Documented
 public @interface Positive {